requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
google-genai>=1.0.0
freewili>=0.1.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
